fix: use new generated files + renaming + type imports in components
